refactor: 모델에서 컨트롤러로 이동
